+++ src/ESP8266_AT_WebServer.h
@@ -11,7 +11,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.5.2
+  Version: 1.5.3
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -39,18 +39,19 @@
   1.5.0   K Hoang      19/12/2021 Reduce usage of Arduino String with std::string
   1.5.1   K Hoang      24/12/2021 Fix bug
   1.5.2   K Hoang      28/12/2021 Fix wrong http status header bug
+  1.5.3   K Hoang      12/01/2022 Fix authenticate issue caused by libb64
  *****************************************************************************************************************************/
 
 #ifndef ESP8266_AT_WebServer_h
 #define ESP8266_AT_WebServer_h
 
-#define ESP8266_AT_WEBSERVER_VERSION          "ESP8266_AT_WebServer v1.5.2"
+#define ESP8266_AT_WEBSERVER_VERSION          "ESP8266_AT_WebServer v1.5.3"
 
 #define ESP8266_AT_WEBSERVER_VERSION_MAJOR    1
 #define ESP8266_AT_WEBSERVER_VERSION_MINOR    5
-#define ESP8266_AT_WEBSERVER_VERSION_PATCH    2
+#define ESP8266_AT_WEBSERVER_VERSION_PATCH    3
 
-#define ESP8266_AT_WEBSERVER_VERSION_INT      1005002
+#define ESP8266_AT_WEBSERVER_VERSION_INT      1005003
 
 #define USE_NEW_WEBSERVER_VERSION         true