+++ src/libb64/cdecode.h
@@ -11,7 +11,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.5.2
+  Version: 1.5.3
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -39,6 +39,7 @@
   1.5.0   K Hoang      19/12/2021 Reduce usage of Arduino String with std::string
   1.5.1   K Hoang      24/12/2021 Fix bug
   1.5.2   K Hoang      28/12/2021 Fix wrong http status header bug
+  1.5.3   K Hoang      12/01/2022 Fix authenticate issue caused by libb64
  ***************************************************************************************************************************************/
 
 #pragma once
@@ -51,15 +52,12 @@
 #define base64_decode_expected_len(n) ((n * 3) / 4)
 
 #ifdef __cplusplus
-  extern "C" {
+extern "C" {
 #endif
 
 typedef enum 
 {
-  step_a, 
-  step_b, 
-  step_c, 
-  step_d
+  step_a, step_b, step_c, step_d
 } base64_decodestep;
 
 typedef struct 
@@ -70,14 +68,14 @@ typedef struct
 
 void base64_init_decodestate(base64_decodestate* state_in);
 
-int base64_decode_value(char value_in);
+int base64_decode_value(int value_in);
 
 int base64_decode_block(const char* code_in, const int length_in, char* plaintext_out, base64_decodestate* state_in);
 
 int base64_decode_chars(const char* code_in, const int length_in, char* plaintext_out);
 
 #ifdef __cplusplus
-  } // extern "C"
+} // extern "C"
 #endif
 
 #endif /* BASE64_CDECODE_H */