+++ src/libb64/cencode.h
@@ -11,7 +11,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.5.2
+  Version: 1.5.3
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -39,6 +39,7 @@
   1.5.0   K Hoang      19/12/2021 Reduce usage of Arduino String with std::string
   1.5.1   K Hoang      24/12/2021 Fix bug
   1.5.2   K Hoang      28/12/2021 Fix wrong http status header bug
+  1.5.3   K Hoang      12/01/2022 Fix authenticate issue caused by libb64
  ***************************************************************************************************************************************/
 
 #pragma once
@@ -48,22 +49,15 @@
 #ifndef BASE64_CENCODE_H
 #define BASE64_CENCODE_H
 
-#define BASE64_CHARS_PER_LINE 72
-
-#define base64_encode_expected_len_nonewlines(n) ((((4 * (n)) / 3) + 3) & ~3)
-#define base64_encode_expected_len(n) \
-       (base64_encode_expected_len_nonewlines(n) + ((n / ((BASE64_CHARS_PER_LINE * 3) / 4)) + 1))
-
+#define base64_encode_expected_len(n) ((((4 * n) / 3) + 3) & ~3)
 
 #ifdef __cplusplus
-  extern "C" {
+extern "C" {
 #endif
 
 typedef enum 
 {
-  step_A, 
-  step_B, 
-  step_C
+  step_A, step_B, step_C
 } base64_encodestep;
 
 typedef struct 
@@ -71,11 +65,9 @@ typedef struct
   base64_encodestep step;
   char result;
   int stepcount;
-  int stepsnewline;
 } base64_encodestate;
 
 void base64_init_encodestate(base64_encodestate* state_in);
-void base64_init_encodestate_nonewlines(base64_encodestate* state_in);
 
 char base64_encode_value(char value_in);
 
@@ -86,7 +78,7 @@ int base64_encode_blockend(char* code_out, base64_encodestate* state_in);
 int base64_encode_chars(const char* plaintext_in, int length_in, char* code_out);
 
 #ifdef __cplusplus
-  } // extern "C"
+} // extern "C"
 #endif
 
 #endif /* BASE64_CENCODE_H */