+++ changelog.md
@@ -12,6 +12,7 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+  * [Releases v1.5.3](#releases-v153)
   * [Releases v1.5.2](#releases-v152)
   * [Releases v1.5.1](#releases-v151)
   * [Releases v1.5.0](#releases-v150)
@@ -38,6 +39,11 @@
 ---
 
 ## Changelog
+
+### Releases v1.5.3
+
+1. Fix authenticate issue caused by libb64
+2. Update `Packages' Patches`
 
 ### Releases v1.5.2